Basic stroke input
The StrokeReader keeps track of the number of characters that have been
received (at least ascii characters), and tries to line that up with the
backspaces that are sent by plover when '*' is pressed.  '*' seems
somewhat hardcoded, and there doesn't seem to be a way to make it just
send the '*'.