Enable stroke test
With the default test profile options, the stroke test takes a long time
to run.  If we enable optimizations during the tests, this test only
takes a few seconds.  Remove the check for an environment variable, and
enable optimizations.  This will run the test every time cargo test is
used.