Add a status display showing new word count
Display a count of how many new words have been added.  This can be used
to decide to stop learning to avoid having too much review burden later.

Note that the display increments when showing a new word, so, for
example, on a new word, if the status area shows 6 words have been
learned, 5 have been added to the learning database.  It is safe to exit
before writing this word, and only the 5 words will be learned.