Schema: Correct constraint on lesson table
The current constraint preventing repeated words in a lesson keeps the
lessons from being usable for regular English prose.  Instead of this
constraint, make a constraint that all of the entries in a given list
must have unique sequence numbers.

This requires a schema migration, which can be found in the migrations
directory.