Fix new word computation
The intent of the code selecting new words was to balance the lists such
that they would be taken from appropriately, ideally to all reach the
end around the same time.

However, the math was backward, and it actually was given priority to
those lists that were already close to completion.

Reverse this priority to give new lists precedence.  The ratio is still
biased a bit toward smaller lists.