ui refactor: Make App a trait
Make the use of the App into a trait within the Ui.  To do this, we
de-generic the render function.  This isn't really a problem, since
we're already storing concrete types for the terminal background in
other places.

Add two methods, a getter and a setter for data fields so that the UI
can use this as a trait.