Record details about errors
Each time an entry has to be corrected, add a database entry capturing
the state of the learn entry at the time, along with the text that the
user actually wrote (including the correction).  This can later be used
for analysis to determine what might be helpful to learning.

Note that this has an accompanied schema change.  A current database can
be migrated with:

  sqlite3 learn.db < migrations/2022-03-12a-2022-04-14a.sql