Remove a couple of explicit discards
Remove a few instances of `_ = thing.pop()`, replacing them with just
`thing.pop()`.  The result is the same, and there doesn't seem to be any
warning about discarding the value.  The pattern assignment is unstable
until fairly recent rustc versions, and fixing this allows us to build
with older versions of Rust.