Try to make learning better with backlog
Currently, having a large backlog means these past entries must be
worked through before fresh words are reviewed.  This makes learning not
as efficient when there are numerous reviews to do.

Help this by changing the order that we examing reviews.  Instead of
just sorting by when they are due (next), sort them first by interval,
so that fresh words get priority over older words.  The main result is
that instead of review being a separate phase before learning can
continue, learning is given priority, and the reviews come in to fill.

One consequence of this is that we can't accurately show upcoming words.
For now, only show one word at a time, to avoid the word list changing
as it approaches.