Change how intervals work, part 1
Two changes to intervals.  When a word is incorrect, don't just reset to
the beginning, as this is just way too costly.  Also account for the
actual interval taken, not just the desired one.  If time passes an the
user still gets it right, base off of that interval.  If it is later
forgotten, the interval will begin to reduce.

The idea is that once something is learned, it is just extra work to set
the interval back to 5 seconds, and we spend most of our time reviewing
these words.  Instead, increase the interval more slowly (averaging
1.5x) but don't punish as much for old words.