Use two phased interval
When computing a new interval, use the last interval time when the
interval is less than a day, and if longer, use the actual time to
compute the new interval.

This forces early learning to have an appropriate number of repeats,
even if learning is infrequent, but allows for words mostly already
learned to have longer intervals if practice time is skipped.  When
practicing regularly, the difference is insignificant, but when time is
skipped, words already known well will have their intervals increase
more.